OWGeneSetsEnrichemnt: merge GeneSets and SetEnrichment in a single widget, bug fix
