+++ orangecontrib/bioinformatics/widgets/OWGeneSetEnrichment.py
@@ -2,23 +2,22 @@
 import threading
 import concurrent.futures
 
-
 from functools import partial
 from collections import defaultdict
 
 from AnyQt.QtWidgets import (
-    QTreeView, QTreeWidget, QTreeWidgetItem, QTreeWidgetItemIterator, QHeaderView
+    QTreeView, QTreeWidget, QTreeWidgetItem, QTreeWidgetItemIterator, QHeaderView, QHBoxLayout
 )
 from AnyQt.QtCore import (
-    Qt, QSize, QThreadPool, QSortFilterProxyModel, Slot, QThread
+    Qt, QSize, QThreadPool, Slot, QThread
 )
 from AnyQt.QtGui import (
     QColor, QStandardItemModel, QStandardItem,
 )
 
 from Orange.widgets.gui import (
-    vBox, lineEdit, ProgressBar, LinkRole, LinkStyledItemDelegate,
-    auto_commit, widgetLabel, spin, comboBox, widgetBox
+    vBox, lineEdit, ProgressBar, LinkRole, LinkStyledItemDelegate, doubleSpin,
+    auto_commit, widgetLabel, spin, comboBox, widgetBox, radioButtonsInBox
 )
 from Orange.data import Domain, Table, DiscreteVariable, StringVariable
 from Orange.widgets.widget import OWWidget, Msg
@@ -32,9 +31,10 @@ from orangecontrib.bioinformatics.widgets.utils.data import (
     ERROR_ON_MISSING_ANNOTATION, ERROR_ON_MISSING_GENE_ID, ERROR_ON_MISSING_TAX_ID
 )
 
-from orangecontrib.bioinformatics.widgets.utils.gui import NumericalColumnDelegate
+from orangecontrib.bioinformatics.widgets.utils.gui import NumericalColumnDelegate, FilterProxyModel
 from orangecontrib.bioinformatics.widgets.utils.concurrent import Worker
 from orangecontrib.bioinformatics.widgets.utils.settings import OrganismContextHandler
+from orangecontrib.bioinformatics.utils.statistics import FDR
 from orangecontrib.bioinformatics.utils import serverfiles
 from orangecontrib.bioinformatics import geneset
 
@@ -75,7 +75,7 @@ class Task:
 
 
 class OWGeneSets(OWWidget):
-    name = "Gene Sets"
+    name = "Gene Set Enrichment"
     description = ""
     icon = "icons/OWGeneSets.svg"
     priority = 9
@@ -86,21 +86,34 @@ class OWGeneSets(OWWidget):
     auto_commit = Setting(True)
     stored_selections = ContextSetting([])
     organism = ContextSetting(None)
-    matched_treshold = Setting(5)
 
-    GENES, COUNT, CATEGORY, TERM,  = range(4)
-    DATA_HEADER_LABELS = ['Genes In Set', 'Count', 'Category', 'Term']
-    HIERARCHY_HEADER_LABELS = ['Category']
+    min_count = Setting(5)
+    use_min_count = Setting(True)
+
+    max_p_value = Setting(0.0001)
+    use_p_value = Setting(False)
+
+    max_fdr = Setting(0.01)
+    use_max_fdr = Setting(True)
+
+    use_reference_data = Setting(True)
+
+    COUNT, REFERENCE, P_VAL, FDR, ENRICHMENT, GENES, CATEGORY, TERM = range(8)
+    DATA_HEADER_LABELS = ["Count", 'Reference', 'p-Value', 'FDR', 'Enrichment', 'Genes In Set', 'Category', 'Term']
 
     class Inputs:
         genes = Input("Genes", Table)
         custom_sets = Input('Custom Gene Sets', Table)
+        reference = Input("Reference Genes", Table)
 
     class Outputs:
         matched_genes = Output("Matched Genes", Table)
 
     class Information(OWWidget.Information):
         pass
+
+    class Warning(OWWidget.Warning):
+        all_sets_filtered = Msg('All sets were filtered out.')
 
     class Error(OWWidget.Error):
         organism_mismatch = Msg('Organism in input data and custom gene sets does not match')
@@ -141,11 +154,22 @@ class OWGeneSets(OWWidget):
         self.custom_gene_id_attribute = None
         self.custom_gene_id_column = None
 
+        # reference genes
+        self.reference_radio_box = None
+        self.reference_data = None
+        self.reference_genes = None
+
+        self.reference_tax_id = None
+        self.reference_attr_names = None
+        self.reference_gene_id_attribute = None
+        self.reference_gene_id_column = None
+
         # info box
         self.input_info = None
         self.num_of_sel_genes = 0
 
         # filter
+        self.line_edit_filter = None
         self.search_pattern = ''
         self.organism_select_combobox = None
 
@@ -181,6 +205,8 @@ class OWGeneSets(OWWidget):
         self.hierarchy_widget.clear()
         # clear data view
         self.init_item_model()
+        # reset filters
+        self.setup_filter_model()
 
     def cancel(self):
         """
@@ -239,7 +265,10 @@ class OWGeneSets(OWWidget):
         self.set_selected_hierarchies()
 
     def update_tree_view(self):
-        self.init_gene_sets()
+        if self.use_reference_data and self.reference_data:
+            self.init_gene_sets(reference_genes=self.reference_genes)
+        else:
+            self.init_gene_sets()
 
     def invalidate(self):
         # clear
@@ -261,6 +290,47 @@ class OWGeneSets(OWWidget):
         if self.tax_id is not None and self.custom_tax_id is not None:
             return self.tax_id != self.custom_tax_id
         return False
+
+    def __get_reference_genes(self):
+        self.reference_genes = []
+
+        if self.reference_attr_names:
+            for variable in self.reference_data.domain.attributes:
+                self.reference_genes.append(str(variable.attributes.get(self.reference_gene_id_attribute, '?')))
+        else:
+            genes, _ = self.reference_data.get_column_view(self.reference_gene_id_column)
+            self.reference_genes = [str(g) for g in genes]
+
+    @Inputs.reference
+    def handle_reference_genes(self, data):
+        """
+        Set the (optional) input dataset with reference gene names.
+        """
+
+        if data:
+            self.reference_data = data
+            self.reference_tax_id = str(self.reference_data.attributes.get(TAX_ID, None))
+            self.reference_attr_names = self.reference_data.attributes.get(GENE_AS_ATTRIBUTE_NAME, None)
+            self.reference_gene_id_attribute = self.reference_data.attributes.get(GENE_ID_ATTRIBUTE, None)
+            self.reference_gene_id_column = self.reference_data.attributes.get(GENE_ID_COLUMN, None)
+
+            if not (self.reference_attr_names is not None
+                    and ((self.reference_gene_id_attribute is None) ^ (self.reference_gene_id_column is None))):
+
+                if self.reference_tax_id is None:
+                    self.Error.missing_annotation()
+                    return
+
+                self.Error.missing_gene_id()
+                return
+
+            elif self.reference_tax_id is None:
+                self.Error.missing_tax_id()
+                return
+
+        self.__get_reference_genes()
+        self.reference_radio_box.setEnabled(bool(self.reference_data))
+        self.invalidate()
 
     @Inputs.custom_sets
     def handle_custom_input(self, data):
@@ -453,8 +523,8 @@ class OWGeneSets(OWWidget):
                     self.gene_sets_obj,
                     sets_to_display,
                     set(self.input_genes),
-                    self.matched_treshold,
                     reference_genes,
+                    self.min_count if self.use_min_count else 1,
                     callback=callback)
 
         progress_iterations = sum([len(g_set) for hier, g_set in self.gene_sets_obj.map_hierarchy_to_sets().items()
@@ -485,6 +555,8 @@ class OWGeneSets(OWWidget):
             results = f.result()  # type: list
             [self.data_model.appendRow(model_item) for model_item in results]
             self.filter_proxy_model.setSourceModel(self.data_model)
+            self._update_fdr()
+            self.filter_data_view()
         except Exception as ex:
             print(ex)
 
@@ -540,6 +612,76 @@ class OWGeneSets(OWWidget):
 
         return sets_to_display
 
+    def filter_data_view(self):
+        filter_proxy = self.filter_proxy_model  # type: FilterProxyModel
+        model = filter_proxy.sourceModel()      # type: QStandardItemModel
+
+        assert isinstance(model, QStandardItemModel)
+
+        search_term = self.search_pattern.lower().strip().split()
+
+        # apply filtering rules
+        filters = [
+            FilterProxyModel.Filter(
+                self.TERM, Qt.DisplayRole,
+                lambda value: all(fs in value.lower() for fs in search_term))
+        ]
+
+        # if self.use_min_count:
+        #    filters.append(
+        #        FilterProxyModel.Filter(
+        #            self.COUNT, Qt.DisplayRole,
+        #            lambda value: value >= self.min_count,
+        #        )
+        #    )
+
+        if self.use_p_value:
+            filters.append(
+                FilterProxyModel.Filter(
+                    self.P_VAL, Qt.DisplayRole,
+                    lambda value: value < self.max_p_value
+                )
+            )
+
+        if self.use_max_fdr:
+            filters.append(
+                FilterProxyModel.Filter(
+                    self.FDR, Qt.DisplayRole,
+                    lambda value: value < self.max_fdr
+                )
+            )
+
+        filter_proxy.set_filters(filters)
+
+        if model.rowCount() and not filter_proxy.rowCount():
+            self.Warning.all_sets_filtered()
+        else:
+            self.Warning.clear()
+
+    def __get_source_data(self, proxy_row_index, column):
+        proxy_index = self.filter_proxy_model.index(proxy_row_index, column)
+        source_index = self.filter_proxy_model.mapToSource(proxy_index)
+        return source_index.data(role=Qt.DisplayRole)
+
+    def _update_fdr(self):
+        # Update the FDR in place due to a changed selected categories set and
+        # results for all of these categories are already available.
+        proxy = self.filter_proxy_model
+        model = self.filter_proxy_model.sourceModel()
+
+        if model is not None:
+            assert isinstance(model, QStandardItemModel)
+
+            p_values = [(i, self.__get_source_data(i, self.P_VAL)) for i in range(proxy.rowCount())]
+            fdr_values = FDR([p_val for _, p_val in p_values])
+
+            for i, fdr_val in zip([i for i, _ in p_values], fdr_values):
+                proxy_index = proxy.index(i, self.FDR)
+                source_index = self.filter_proxy_model.mapToSource(proxy_index)
+                source_item = model.item(source_index.row(), self.FDR)
+                source_item.setData(fdr_val, role=Qt.DisplayRole)
+                source_item.setData(fdr_val, role=Qt.ToolTipRole)
+
     def commit(self):
         selection_model = self.data_view.selectionModel()
 
@@ -585,32 +727,82 @@ class OWGeneSets(OWWidget):
             self.COUNT, NumericalColumnDelegate(self)
         )
 
+        self.data_view.setItemDelegateForColumn(
+            self.REFERENCE, NumericalColumnDelegate(self)
+        )
+
+        self.data_view.setItemDelegateForColumn(
+            self.P_VAL, NumericalColumnDelegate(self, precision=2, notation='e')
+        )
+
+        self.data_view.setItemDelegateForColumn(
+            self.FDR, NumericalColumnDelegate(self, precision=2, notation='e')
+        )
+
+        self.data_view.setItemDelegateForColumn(
+            self.ENRICHMENT, NumericalColumnDelegate(self, precision=1)
+        )
+
     def setup_filter_model(self):
-        self.filter_proxy_model = QSortFilterProxyModel()
+        self.filter_proxy_model = FilterProxyModel()
         self.filter_proxy_model.setFilterKeyColumn(self.TERM)
+        self.data_view.setModel(self.filter_proxy_model)
 
     def setup_filter_area(self):
-        line_edit_filter = lineEdit(self.mainArea, self, 'search_pattern', 'Filter gene sets:')
-        line_edit_filter.setPlaceholderText('search pattern ...')
-        line_edit_filter.textChanged.connect(self.filter_proxy_model.setFilterRegExp)
+        h_layout = QHBoxLayout()
+        h_layout.setSpacing(100)
+        h_widget = widgetBox(self.mainArea, orientation=h_layout)
+
+        spin(h_widget, self, 'min_count', 0, 100,
+             label='Count',
+             tooltip='Minimum genes count',
+             checked='use_min_count',
+             callback=self.invalidate,
+             callbackOnReturn=True,
+             checkCallback=self.invalidate)
+
+        doubleSpin(h_widget, self, 'max_p_value', 0.0, 1.0, 0.0001,
+                   label='p-value',
+                   tooltip='Maximum p-value of the enrichment score',
+                   checked='use_p_value',
+                   callback=self.filter_data_view,
+                   callbackOnReturn=True,
+                   checkCallback=self.filter_data_view
+                   )
+
+        doubleSpin(h_widget, self, 'max_fdr', 0.0, 1.0, 0.0001,
+                   label='FDR',
+                   tooltip='Maximum false discovery rate',
+                   checked='use_max_fdr',
+                   callback=self.filter_data_view,
+                   callbackOnReturn=True,
+                   checkCallback=self.filter_data_view
+                   )
+
+        self.line_edit_filter = lineEdit(h_widget, self, 'search_pattern')
+        self.line_edit_filter.setPlaceholderText('Filter gene sets ...')
+        self.line_edit_filter.textChanged.connect(self.filter_data_view)
 
     def setup_control_area(self):
         info_box = vBox(self.controlArea, 'Info')
         self.input_info = widgetLabel(info_box)
-
-        box = vBox(self.controlArea, "Minimum Count")
-        self.spin_widget = spin(box, self, 'matched_treshold', 0, 1000,
-                                callback=self.invalidate)
 
         box = vBox(self.controlArea, "Custom Gene Sets")
         self.gs_label_combobox = comboBox(box, self, "gene_set_label", sendSelectedValue=True,
                                           model=self.feature_model, callback=self.invalidate)
         self.gs_label_combobox.setDisabled(True)
 
-        hierarchy_box = widgetBox(self.controlArea, "Gene Set Databases")
+        self.reference_radio_box = radioButtonsInBox(
+            self.controlArea, self, "use_reference_data", ["Entire genome", "Reference gene set (input)"],
+            tooltips=["Use entire genome (for gene set enrichment)", "Use reference set of genes"],
+            box="Reference", callback=self.invalidate)
+
+        self.reference_radio_box.setEnabled(False)
+
+        hierarchy_box = widgetBox(self.controlArea, "Gene Set Categories")
         self.hierarchy_widget = QTreeWidget(self)
         self.hierarchy_widget.setEditTriggers(QTreeView.NoEditTriggers)
-        self.hierarchy_widget.setHeaderLabels(self.HIERARCHY_HEADER_LABELS)
+        self.hierarchy_widget.setHeaderLabels([' '])
         self.hierarchy_widget.itemClicked.connect(self.update_tree_view)
         hierarchy_box.layout().addWidget(self.hierarchy_widget)
 
@@ -621,11 +813,9 @@ class OWGeneSets(OWWidget):
         self.setup_control_area()
 
         # main area
+        self.data_view = QTreeView()
         self.setup_filter_model()
         self.setup_filter_area()
-
-        self.data_view = QTreeView()
-        self.data_view.setModel(self.filter_proxy_model)
         self.data_view.setAlternatingRowColors(True)
         self.data_view.sortByColumn(self.COUNT, Qt.DescendingOrder)
         self.data_view.setSortingEnabled(True)
@@ -641,7 +831,7 @@ class OWGeneSets(OWWidget):
         self.assign_delegates()
 
     @staticmethod
-    def set_items(gene_sets, sets_to_display, genes, matched_treshold, ref, callback):
+    def set_items(gene_sets, sets_to_display, genes, ref, count_treshold, callback):
         model_items = []
         if not genes:
             return
@@ -649,36 +839,47 @@ class OWGeneSets(OWWidget):
         for gene_set in gene_sets:
             if gene_set.hierarchy not in sets_to_display:
                 continue
-
-            category_column = QStandardItem()
-            name_column = QStandardItem()
-            matched_column = QStandardItem()
-            genes_column = QStandardItem()
-
-            category_column.setData(", ".join(gene_set.hierarchy), Qt.DisplayRole)
-            name_column.setData(gene_set.name, Qt.DisplayRole)
-            name_column.setData(gene_set.name, Qt.ToolTipRole)
-            name_column.setData(gene_set.link, LinkRole)
-            name_column.setForeground(QColor(Qt.blue))
-
-            matched_set = gene_set.genes & genes
-            matched_column.setData(matched_set, Qt.UserRole)
-            matched_column.setData(len(matched_set), Qt.DisplayRole)
-
-            genes_column.setData(len(gene_set.genes), Qt.DisplayRole)
-            genes_column.setData(gene_set.genes, Qt.UserRole)  # store genes to get then on output on selection
-
+            enrichemnt_result = gene_set.set_enrichment(ref, genes.intersection(ref))
             callback()
 
-            if len(matched_set) >= matched_treshold:
-                model_items.append([genes_column, matched_column, name_column, category_column])
+            if len(enrichemnt_result.query) >= count_treshold:
+                category_column = QStandardItem()
+                name_column = QStandardItem()
+                count_column = QStandardItem()
+                genes_column = QStandardItem()
+                ref_column = QStandardItem()
+                pval_column = QStandardItem()
+                fdr_column = QStandardItem()
+                enrichemnt_column = QStandardItem()
 
+                category_column.setData(", ".join(gene_set.hierarchy), Qt.DisplayRole)
+                name_column.setData(gene_set.name, Qt.DisplayRole)
+                name_column.setData(gene_set.name, Qt.ToolTipRole)
+                name_column.setData(gene_set.link, LinkRole)
+                name_column.setForeground(QColor(Qt.blue))
+
+                count_column.setData(len(enrichemnt_result.query), Qt.DisplayRole)
+                count_column.setData(set(enrichemnt_result.query), Qt.UserRole)
+
+                genes_column.setData(len(gene_set.genes), Qt.DisplayRole)
+                genes_column.setData(set(gene_set.genes), Qt.UserRole)  # store genes to get then on output on selection
+
+                ref_column.setData(len(enrichemnt_result.reference), Qt.DisplayRole)
+
+                pval_column.setData(enrichemnt_result.p_value, Qt.DisplayRole)
+                pval_column.setData(enrichemnt_result.p_value, Qt.ToolTipRole)
+
+                enrichemnt_column.setData(enrichemnt_result.enrichment_score, Qt.DisplayRole)
+                enrichemnt_column.setData(enrichemnt_result.enrichment_score, Qt.ToolTipRole)
+
+                model_items.append([count_column, ref_column, pval_column, fdr_column, enrichemnt_column,
+                                    genes_column, category_column, name_column])
         return model_items
 
     def init_item_model(self):
         if self.data_model:
             self.data_model.clear()
-            self.filter_proxy_model.setSourceModel(None)
+            self.setup_filter_model()
         else:
             self.data_model = QStandardItemModel()
